fix(memory): load raaf-core as a gem, and unlock the file store
The stores reached into ../../../core/lib by relative path. That works only
from a checkout of the mono-repo laid out exactly this way, and it loads a
second copy of RAAF::Utils beside the one the host application already has.
They require the gem.

FileStore#search called #retrieve while holding @mutex, and #retrieve takes
it again — Ruby's Mutex is not reentrant, so any search over a non-empty
index deadlocked. The read is now a private method both paths share.

Two smaller ones: recent memories sorted by updated_at.to_i, so everything
written within the same second came back in hash order, and the summarizer
wrote Memory::Memory from inside RAAF::Memory, which looks for a Memory
nested in the Memory class.

+++ memory/lib/raaf/memory/in_memory_store.rb
@@ -150,7 +150,7 @@ module RAAF
       def get_recent(limit = 10)
         @mutex.synchronize do
           @memories.values
-                   .sort_by { |m| -m.updated_at.to_i }
+                   .sort_by { |m| -m.updated_at.to_f }
                    .take(limit)
                    .map(&:to_h)
         end
